+++ macos/Classes/FlutterAppUpdaterPlugin.swift
@@ -12,6 +12,10 @@ public class FlutterAppUpdaterPlugin: NSObject, FlutterPlugin {
     switch call.method {
     case "getPlatformVersion":
       result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
+    case "getAppVersionName":
+      result(Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String)
+    case "getAppVersionCode":
+      result(Bundle.main.infoDictionary?["CFBundleVersion"] as? String)
     default:
       result(FlutterMethodNotImplemented)
     }